+++ src/components/PurchaseOrder/Summary/TotalEncumberedValue/TotalEncumberedValue.js
@@ -1,44 +1,14 @@
-import React, {
-  useEffect,
-  useState,
-} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
-import { stripesConnect } from '@folio/stripes/core';
 import {
   KeyValue,
-  Loading,
 } from '@folio/stripes/components';
 import {
   AmountWithCurrencyField,
-  transactionsManifest,
 } from '@folio/stripes-acq-components';
 
-const TotalEncumberedValue = ({ orderId, label, mutator }) => {
-  const [totalEncumbered, setTotalEncumbered] = useState();
-
-  useEffect(
-    () => {
-      setTotalEncumbered();
-
-      if (orderId) {
-        mutator.orderTransactions.GET()
-          .then(transactions => {
-            const total = transactions.reduce((acc, { amount }) => acc + amount, 0);
-
-            setTotalEncumbered(total);
-          })
-          .catch(() => setTotalEncumbered(0));
-      }
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [orderId],
-  );
-
-  if (totalEncumbered === undefined) {
-    return <Loading />;
-  }
-
+const TotalEncumberedValue = ({ totalEncumbered, label }) => {
   return (
     <KeyValue label={label}>
       <AmountWithCurrencyField amount={totalEncumbered} />
@@ -46,19 +16,9 @@ const TotalEncumberedValue = ({ orderId, label, mutator }) => {
   );
 };
 
-TotalEncumberedValue.manifest = Object.freeze({
-  orderTransactions: {
-    ...transactionsManifest,
-    params: {
-      query: 'encumbrance.sourcePurchaseOrderId==!{orderId}',
-    },
-  },
-});
-
 TotalEncumberedValue.propTypes = {
-  orderId: PropTypes.string,
+  totalEncumbered: PropTypes.number,
   label: PropTypes.node.isRequired,
-  mutator: PropTypes.object.isRequired,
 };
 
-export default stripesConnect(TotalEncumberedValue);
+export default TotalEncumberedValue;